refactor: split lib and cli files to speed up cli
