+++ aimbat/app.py
@@ -1,19 +1,18 @@
-from aimbat.lib import (
-    project,
-    defaults,
+from aimbat.cli import (
     checkdata,
-    sampledata,
     data,
-    station,
-    seismogram,
+    defaults,
     event,
+    project,
+    sampledata,
+    seismogram,
     snapshot,
+    station,
     utils,
 )
-
-# from aimbat.gui import aimbat_main_gui
 from importlib import metadata
 import click
+
 
 try:
     __version__ = metadata.version("aimbat")
@@ -48,15 +47,15 @@ def cli(ctx: click.Context, debug: bool, use_qt: bool) -> None:
 #     aimbat_main_gui()
 
 
-cli.add_command(project.project_cli)
-cli.add_command(defaults.defaults_cli)
-cli.add_command(sampledata.sampledata_cli)
 cli.add_command(checkdata.checkdata_cli)
 cli.add_command(data.data_cli)
-cli.add_command(station.station_cli)
+cli.add_command(defaults.defaults_cli)
 cli.add_command(event.event_cli)
+cli.add_command(project.project_cli)
+cli.add_command(sampledata.sampledata_cli)
 cli.add_command(seismogram.seismogram_cli)
 cli.add_command(snapshot.snapshot_cli)
+cli.add_command(station.station_cli)
 cli.add_command(utils.utils_cli)
