+++ aimbat/lib/data.py
@@ -1,6 +1,6 @@
 """Module to add seismogram files to an AIMBAT project and view information about them."""
 
-from datetime import timedelta
+from aimbat.lib.common import ic
 from aimbat.lib.defaults import defaults_get_value
 from aimbat.lib.models import (
     AimbatFile,
@@ -12,17 +12,14 @@ from aimbat.lib.models import (
     AimbatSeismogramParameter,
 )
 from aimbat.lib.db import engine
-from aimbat.lib.common import AIMBAT_FILE_TYPES, AimbatFileType, cli_enable_debug
+from aimbat.lib.types import AimbatFileType
 from aimbat.lib.io import read_metadata_from_file
+from datetime import timedelta
 from pathlib import Path
 from sqlmodel import Session, select
 from rich.progress import track
 from rich.console import Console
 from rich.table import Table
-from icecream import ic  # type: ignore
-import click
-
-ic.disable()
 
 
 def data_add_files(
@@ -181,33 +178,3 @@ def data_print_table() -> None:
 
     console = Console()
     console.print(table)
-
-
-@click.group("data")
-@click.pass_context
-def data_cli(ctx: click.Context) -> None:
-    """Manage data in the AIMBAT project."""
-    cli_enable_debug(ctx)
-
-
-@data_cli.command("add")
-@click.option(
-    "--filetype",
-    type=click.Choice(AIMBAT_FILE_TYPES, case_sensitive=False),
-    default="sac",
-    help="File type.",
-)
-@click.argument("data_files", nargs=-1, type=click.Path(exists=True), required=True)
-def cli_add(data_files: list[Path], filetype: AimbatFileType) -> None:
-    """Add or update data files in the AIMBAT project."""
-    data_add_files(data_files, filetype, disable_progress_bar=ic.enabled)
-
-
-@data_cli.command("list")
-def cli_list() -> None:
-    """Print information on the data stored in AIMBAT."""
-    data_print_table()
-
-
-if __name__ == "__main__":
-    data_cli(obj={})
