+++ aimbat/lib/project.py
@@ -1,3 +1,4 @@
+from aimbat.lib.common import ic
 from aimbat.lib.db import engine, AIMBAT_PROJECT
 from aimbat.lib.models import (
     AimbatEvent,
@@ -7,16 +8,11 @@ from aimbat.lib.models import (
     AimbatStation,
 )
 from aimbat.lib.defaults import defaults_load_global_values
-from aimbat.lib.common import cli_enable_debug
 from sqlmodel import SQLModel, Session, select, text
 from pathlib import Path
 from typing import Any
-from icecream import ic  # type: ignore
 from rich.console import Console
-import click
 import aimbat.lib.models  # noqa: F401
-
-ic.disable()
 
 
 def project_exists() -> bool:
@@ -101,51 +97,3 @@ def project_print_info() -> Any:
             "Number of Seismograms (total/selected): ",
             f"({len(all_seismograms)}/{len(selected_seismograms)})",
         )
-
-
-@click.group("project")
-@click.pass_context
-def project_cli(ctx: click.Context) -> None:
-    """Manage an AIMBAT project file.
-
-    This command manages an AIMBAT project. By default, the project consists
-    of a file called `aimbat.db` in the current working directory. All AIMBAT
-    commands must be executed from the same directory.
-
-    The location (and name) of the project file may also be specified by
-    setting the AIMBAT_PROJECT environment variable to the desired filename.
-    """
-    cli_enable_debug(ctx)
-
-
-@project_cli.command("new")
-def project_cli_new() -> None:
-    """Create new AIMBAT project."""
-    try:
-        project_new()
-        print(f"Created new AIMBAT project in {AIMBAT_PROJECT}.")
-    except FileExistsError as e:
-        print(e)
-
-
-@project_cli.command("del")
-@click.confirmation_option(prompt="Are you sure?")
-def project_cli_del() -> None:
-    """Delete project (note: this does not delete seismogram data)."""
-    try:
-        project_del()
-    except FileNotFoundError as e:
-        print(e)
-
-
-@project_cli.command("info")
-def project_cli_info() -> None:
-    """Show information on an exisiting project."""
-    try:
-        project_print_info()
-    except FileNotFoundError as e:
-        print(e)
-
-
-if __name__ == "__main__":
-    project_cli(obj={})
