+++ aimbat/lib/seismogram.py
@@ -1,22 +1,13 @@
+from aimbat.lib.common import ic
 from aimbat.lib.db import engine
 from aimbat.lib.models import AimbatSeismogram
-from aimbat.lib.common import RegexEqual, cli_enable_debug
+from aimbat.lib.types import (
+    AimbatSeismogramParameterName,
+    AimbatSeismogramParameterType,
+)
 from rich.console import Console
 from rich.table import Table
 from sqlmodel import Session, select
-from datetime import datetime, timedelta
-from typing import Literal, Tuple, get_args
-from icecream import ic  # type: ignore
-import click
-
-ic.disable()
-
-# Valid AIMBAT parameter types and names
-AimbatSeismogramParameterType = float | datetime | timedelta | str
-AimbatSeismogramParameterName = Literal["select", "t1", "t2"]
-AIMBAT_SEISMOGRAM_PARAMETER_NAMES: Tuple[AimbatSeismogramParameterName, ...] = get_args(
-    AimbatSeismogramParameterName
-)
 
 
 def seismogram_get_parameter(
@@ -101,73 +92,3 @@ def seismogram_print_table() -> None:
 
     console = Console()
     console.print(table)
-
-
-@click.group("seismogram")
-@click.pass_context
-def seismogram_cli(ctx: click.Context) -> None:
-    """View and manage seismograms in the AIMBAT project."""
-    cli_enable_debug(ctx)
-
-
-@seismogram_cli.command("list")
-def seismogram_cli_list() -> None:
-    """Print information on the seismograms stored in AIMBAT."""
-    seismogram_print_table()
-
-
-@seismogram_cli.group("parameter")
-def seismogram_cli_parameter() -> None:
-    """Manage evennt parameters in the AIMBAT project."""
-    pass
-
-
-@seismogram_cli_parameter.command("get")
-@click.argument("seismogram_id", nargs=1, type=int, required=True)
-@click.argument(
-    "name", nargs=1, type=click.Choice(AIMBAT_SEISMOGRAM_PARAMETER_NAMES), required=True
-)
-def seismogram_cli_paramter_get(
-    seismogram_id: int, name: AimbatSeismogramParameterName
-) -> None:
-    """Get the value of a processing parameter."""
-
-    with Session(engine) as session:
-        print(seismogram_get_parameter(session, seismogram_id, name))
-
-
-@seismogram_cli_parameter.command("set")
-@click.argument(
-    "seismogram_id",
-    nargs=1,
-    type=int,
-    required=True,
-)
-@click.argument(
-    "name", nargs=1, type=click.Choice(AIMBAT_SEISMOGRAM_PARAMETER_NAMES), required=True
-)
-@click.argument("value", nargs=1, type=str, required=True)
-def seismogram_cli_parameter_set(
-    seismogram_id: int,
-    name: AimbatSeismogramParameterName,
-    value: str,
-) -> None:
-    """Set value of a processing parameter."""
-
-    with Session(engine) as session:
-        match [name, RegexEqual(value)]:
-            case ["select", "True" | "true" | "yes" | "y"]:
-                seismogram_set_parameter(session, seismogram_id, name, True)
-            case ["select", "False" | "false" | "no" | "n"]:
-                seismogram_set_parameter(session, seismogram_id, name, False)
-            case ["t1" | "t2", r"\d\d\d\d[W,T,0-9,\-,:,\.,\s]+"]:
-                datetime_object = datetime.fromisoformat(value)
-                seismogram_set_parameter(session, seismogram_id, name, datetime_object)
-            case _:
-                raise RuntimeError(
-                    f"Unknown parameter name '{name}' or incorrect value '{value}'."
-                )
-
-
-if __name__ == "__main__":
-    seismogram_cli(obj={})
