+++ aimbat/lib/utils.py
@@ -1,16 +1,12 @@
 from aimbat.lib.db import engine
 from aimbat.lib.models import AimbatSeismogram
-from aimbat.lib.common import cli_enable_debug, check_for_notebook
+from aimbat.lib.common import check_for_notebook, ic
 from pysmo import MiniSeismogram, distance, time_array, unix_time_array
 from sqlmodel import Session, select
-from icecream import ic  # type: ignore
 from pyqtgraph.jupyter import PlotWidget  # type: ignore
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 import pyqtgraph as pg  # type: ignore
-import click
-
-ic.disable()
 
 
 def utils_plotseis(event_id: int, use_qt: bool = False) -> None | PlotWidget:
@@ -82,31 +78,3 @@ def utils_plotseis(event_id: int, use_qt: bool = False) -> None | PlotWidget:
             plt.title(title)
             plt.show()
         return None
-
-
-@click.group("utils")
-@click.pass_context
-def utils_cli(ctx: click.Context) -> None:
-    """Usefull extra tools to use with AIMBAT projects."""
-    cli_enable_debug(ctx)
-
-
-@utils_cli.command("plotseis")
-@click.argument("event_id", nargs=1, type=int, required=True)
-@click.pass_context
-def utils_cli_plotseis(ctx: click.Context, event_id: int) -> None:
-    """Plot seismograms in project."""
-
-    use_qt: bool = ctx.obj.get("USE_QT", False)
-
-    if use_qt:
-        pg.mkQApp()
-
-    utils_plotseis(event_id, use_qt)
-
-    if use_qt:
-        pg.exec()
-
-
-if __name__ == "__main__":
-    utils_cli(obj={})
